requirements.txt
sqlalchemy~=2.0.28
psycopg==3.2.1
sqlmodel==0.0.20
fastapi~=0.104.1
uvicorn~=0.24.0.post1
structlog~=23.2.0
pydantic-settings~=2.1.0
httpx